refactor referenced object passing to separate files
