+++ src/dataStructures.h
@@ -6,11 +6,25 @@
 //
 //
 
+#pragma once
+
 #ifndef dataStructures_h
 #define dataStructures_h
 
 
 #endif /* dataStructures_h */
+
+
+struct volca { // central volca object
+    bool recording;
+    bool playing;
+    bool paused;
+    bool singleShot;
+    int recordFPS;
+    int recordWidth, recordHeight, recordStep;
+    string recordingDate;
+};
+
 
 struct vRenderer { // rendering data for volca object
     bool showGui;
@@ -24,3 +38,4 @@ struct vRenderer { // rendering data for volca object
     float depthFactor;
     float perspectiveFactor;
 } ;
+
