+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include "ofApp.h"
-
+#include "dataStructures.h"
 int main() {
     ofSetupOpenGL(1080, 800, OF_WINDOW);
